requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
requests==2.31.0
pydantic==2.5.1
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn==21.2.0